Fix some CRC issues, timeouts, and other things

+++ src/emulation/wireless/crc.h
@@ -21,7 +21,7 @@
 #define TRUE	!FALSE
 #endif
 
-typedef unsigned short  crc;
+typedef BYTE  crc;
 
 #define CRC_NAME			"CRC-CCITT"
 #define POLYNOMIAL			0x107
